Make Identification a subclass of Comment
