+++ pyinaturalist/models/identification.py
@@ -14,14 +14,57 @@ from pyinaturalist.models import (
 
 
 @define_model
-class Identification(BaseModel):
+class Comment(BaseModel):
+    """:fa:`comment` An observation comment, based on the schema of comments
+    from `GET /observations <https://api.inaturalist.org/v1/docs/#!/Observations/get_observations>`_.
+    """
+
+    body: str = field(default='', doc='Comment text')
+    created_at: datetime = datetime_now_field(doc='Date and time the comment was created')
+    hidden: bool = field(default=None, doc='Indicates if the comment is hidden')
+    uuid: str = field(default=None, doc='Universally unique identifier')
+    user: property = LazyProperty(
+        User.from_json, type=User, doc='User that added the comment or ID'
+    )
+
+    # Unused attributes
+    # created_at_details: Dict = field(factory=dict)
+    # flags: List = field(factory=list)
+    # moderator_actions: List = field(factory=list)
+
+    @property
+    def truncated_body(self):
+        """Comment text, truncated"""
+        truncated_body = self.body.replace('\n', ' ').strip()
+        if len(truncated_body) > 50:
+            truncated_body = truncated_body[:47].strip() + '...'
+        return truncated_body
+
+    @property
+    def username(self) -> str:
+        return self.user.login
+
+    @property
+    def _row(self) -> TableRow:
+        return {
+            'ID': self.id,
+            'User': self.username,
+            'Created at': self.created_at,
+            'Comment': self.truncated_body,
+        }
+
+    @property
+    def _str_attrs(self) -> List[str]:
+        return ['id', 'username', 'created_at', 'truncated_body']
+
+
+@define_model
+class Identification(Comment):
     """:fa:`fingerprint` An observation identification, based on the schema of
     `GET /identifications <https://api.inaturalist.org/v1/docs/#!/Identifications/get_identifications>`_.
     """
 
-    body: str = field(default=None, doc='Comment text')
     category: str = field(default=None, options=ID_CATEGORIES, doc='Identification category')
-    created_at: datetime = datetime_now_field(doc='Date and time the identification was added')
     current: bool = field(
         default=None, doc='Indicates if the identification is the currently accepted one'
     )
@@ -29,21 +72,15 @@ class Identification(BaseModel):
     disagreement: bool = field(
         default=None, doc='Indicates if this identification disagrees with previous ones'
     )
-    hidden: bool = field(default=None)
     own_observation: bool = field(
         default=None, doc='Indicates if the indentifier is also the observer'
     )
     previous_observation_taxon_id: int = field(default=None, doc='Previous observation taxon ID')
-    taxon_change: bool = field(default=None)  # TODO: confirm type
-    taxon_id: int = field(default=None, doc='Identification taxon ID')
-    uuid: str = field(default=None, doc='Universally unique identifier')
+    taxon_change: bool = field(default=None)
     vision: bool = field(
         default=None, doc='Indicates if the taxon was selected from computer vision results'
     )
     taxon: property = LazyProperty(Taxon.from_json, type=Taxon, doc='Identification taxon')
-    user: property = LazyProperty(
-        User.from_json, type=User, doc='User that added the indentification'
-    )
 
     @property
     def taxon_name(self) -> str:
@@ -55,6 +92,7 @@ class Identification(BaseModel):
     # flags: List = field(factory=list)
     # moderator_actions: List = field(factory=list)
     # observation: {}
+    # taxon_id: int = field(default=None)
 
     @property
     def _row(self) -> TableRow:
@@ -69,4 +107,4 @@ class Identification(BaseModel):
 
     @property
     def _str_attrs(self) -> List[str]:
-        return ['id', 'taxon_name', 'created_at']
+        return ['id', 'username', 'taxon_name', 'created_at', 'truncated_body']
